server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
pinecone-client==3.0.0
groq==0.4.1
psycopg2-binary==2.9.9
